feat(hotel): Implement Hotel CRUD operations and fix inconsistencies
This commit introduces the following changes:

- **Backend:**
  - Restructured the backend to have a dedicated `Hotel` module.
  - Moved the `Hotel` type definition from `Business/typeDefs.ts` to `Hotel/typeDefs.ts`.
  - Moved the `Hotel` resolvers from `Business/resolvers.ts` to `Hotel/resolvers.ts`.
  - Updated the main `index.ts` to import and use the new `hotelResolvers` and `hotelTypeDef`.

- **Frontend:**
  - Created GraphQL documents for `Hotel` CRUD operations (`createHotel`, `updateHotel`, `deleteHotel`, `hotels`, `hotel`).
  - Created GraphQL documents for `Room` CRUD operations (`createRoom`, `updateRoom`, `deleteRoom`, `rooms`).

I was unable to complete the task because I was stuck trying to run the GraphQL Code Generator. The backend server was not starting correctly, and I was unable to generate the necessary types and hooks for the frontend. I tried several approaches to fix this, including updating dependencies, running different scripts, and installing missing packages, but none of them worked.
