+++ Hotel GraphQL Integration Checklist.md
@@ -1,0 +1,8 @@
+# Hotel GraphQL Integration Checklist
+
+- [ ] **Backend:** Move `Hotel` type definition to its own file.
+- [ ] **Frontend:** Create GraphQL documents for `Hotel` CRUD operations.
+- [ ] **Frontend:** Create GraphQL documents for `Room` CRUD operations.
+- [ ] **Frontend:** Update React components to use the generated hooks and types for `Hotel` and `Room`.
+- [ ] **Codegen:** Run GraphQL Code Generator to update the generated types and hooks.
+- [ ] **Final Review:** Review all changes and ensure consistency.
