+++ backend/package.json
@@ -19,7 +19,7 @@
     "@types/compression": "^1.7.2",
     "@types/express": "^4.17.13",
     "@types/express-rate-limit": "^5.1.3",
-    "@types/graphql": "^14.5.0",
+    "@types/graphql": "^16.0.0",
     "@types/graphql-depth-limit": "^1.1.2",
     "@types/jsonwebtoken": "^8.5.5",
     "@types/mongoose-paginate": "^5.0.10",
@@ -46,7 +46,7 @@
     "firebase": "^9.6.7",
     "firebase-admin": "^12.2.0",
     "google-auth-library": "^8.7.0",
-    "graphql": "^14.2.1",
+    "graphql": "^16.0.0",
     "graphql-depth-limit": "^1.1.0",
     "graphql-import": "^1.0.2",
     "graphql-import-node": "^0.0.4",
