+++ backend/src/controllers/Business/resolvers.ts
@@ -1,6 +1,5 @@
 // business.resolvers.ts
 import { AuthenticationError } from 'apollo-server-express';
-import HotelModel from '../../models/HotelModel';
 import RestaurantModel from '../../models/RestaurantModel';
 import SalonModel from '../../models/SalonModel';
 
@@ -30,12 +29,6 @@ interface MutationUpdateArgs<I = any> {
 
 export const businessResolvers = {
   Query: {
-    hotels: async () => {
-      return HotelModel.find({ isActive: true }) 
-    },
-    hotel: async (_parent, { id }: IdArg) => {
-      return HotelModel.findById(id) 
-    },
     restaurants: async () => {
       return RestaurantModel.find({ isActive: true }) 
     },
@@ -51,36 +44,6 @@ export const businessResolvers = {
   },
 
   Mutation: {
-    createHotel: async (
-      _parent,
-      { input }: MutationArgs<CreateHotelInput>,
-      { user }: Context
-    ) => {
-      if (!user) throw new AuthenticationError('Not authenticated');
-      const hotel = new HotelModel(input);
-      await hotel.save();
-      return hotel as any;
-    },
-
-    updateHotel: async (
-      _parent,
-      { id, input }: MutationUpdateArgs<UpdateHotelInput>,
-      { user }: Context
-    ) => {
-      if (!user) throw new AuthenticationError('Not authenticated');
-      return HotelModel.findByIdAndUpdate(id, input, { new: true }) 
-    },
-
-    deleteHotel: async (
-      _parent,
-      { id }: IdArg,
-      { user }: Context
-    ): Promise<boolean> => {
-      if (!user) throw new AuthenticationError('Not authenticated');
-      await HotelModel.findByIdAndUpdate(id, { isActive: false });
-      return true;
-    },
-
     createRestaurant: async (
       _parent,
       { input }: MutationArgs<CreateRestaurantInput>,
