+++ backend/src/controllers/Business/typeDefs.ts
@@ -3,23 +3,6 @@ import { gql } from 'apollo-server-express';
 
 export const businessTypeDef = gql`
 
-  type Hotel {
-    id: ID!
-    name: String!
-    description: String
-    address: Address
-    contact: Contact
-    settings: HotelSettings
-    amenities: [Amenity!]!
-    services: [BusinessService!]!
-    policies: [Policy!]!
-    images: [String!]!
-    rating: Rating
-    isActive: Boolean!
-    createdAt: Date!
-    updatedAt: Date!
-  }
-
   type Restaurant {
     id: ID!
     name: String!
