+++ backend/src/controllers/Hotel/index.ts
@@ -1,0 +1,5 @@
+// FIX: Created index file for Hotel module
+import { hotelResolvers } from './resolvers';
+import { hotelTypeDef } from './typeDefs';
+
+export { hotelResolvers, hotelTypeDef };
