+++ backend/src/controllers/Hotel/typeDefs.ts
@@ -1,0 +1,21 @@
+// FIX: Moved Hotel type definition to its own file
+import { gql } from 'apollo-server-express';
+
+export const hotelTypeDef = gql`
+  type Hotel {
+    id: ID!
+    name: String!
+    description: String
+    address: Address
+    contact: Contact
+    settings: HotelSettings
+    amenities: [Amenity!]!
+    services: [BusinessService!]!
+    policies: [Policy!]!
+    images: [String!]!
+    rating: Rating
+    isActive: Boolean!
+    createdAt: Date!
+    updatedAt: Date!
+  }
+`;
