+++ frontend/src/graphql/mutations/createRoom.graphql
@@ -1,0 +1,20 @@
+# FIX: Added GraphQL mutation for creating a room
+mutation CreateRoom($input: RoomInput!) {
+  createRoom(input: $input) {
+    id
+    hotelId
+    number
+    type
+    floor
+    capacity
+    price
+    size
+    status
+    amenities
+    features
+    condition
+    lastCleaned
+    nextMaintenance
+    images
+  }
+}
