+++ frontend/src/graphql/mutations/updateHotel.graphql
@@ -1,0 +1,20 @@
+# FIX: Added GraphQL mutation for updating a hotel
+mutation UpdateHotel($id: ID!, $input: HotelInput!) {
+  updateHotel(id: $id, input: $input) {
+    id
+    name
+    description
+    address {
+      street
+      city
+      state
+      zipCode
+      country
+    }
+    contact {
+      phone
+      email
+      website
+    }
+  }
+}
