+++ frontend/src/graphql/mutations/updateRoom.graphql
@@ -1,0 +1,20 @@
+# FIX: Added GraphQL mutation for updating a room
+mutation UpdateRoom($id: ID!, $input: RoomInput!) {
+  updateRoom(id: $id, input: $input) {
+    id
+    hotelId
+    number
+    type
+    floor
+    capacity
+    price
+    size
+    status
+    amenities
+    features
+    condition
+    lastCleaned
+    nextMaintenance
+    images
+  }
+}
